Move getXXXKeys to jwt/core/
