+++ packages/jwt/src/index.ts
@@ -6,10 +6,10 @@
 
 export {
   getSecretOrPrivateKey,
-} from './get-secret-or-private-key.util';
+} from './core/get-secret-or-private-key.util';
 export {
   getSecretOrPublicKey,
-} from './get-secret-or-public-key.util';
+} from './core/get-secret-or-public-key.util';
 export {
   InvalidTokenError,
   isInvalidTokenError,
